Ensure only the top most visible base layer is visible
Previously if more than one base layer was visible then the state of the
base layer radio buttons was inconsistent with the layer visibility as
only one radio button can be enabled at a time.

This commit looks to ensure that only the top most base layer is visible
prior to layer list being rendered to ensure consistency.