Fix layer id
In the DOM, an ID is assigned to each layer. The ID is generated by replacing spaces with hyphens.

Currently, the replace function will only find the first space and replace it with a hyphen. All other spaces remain. This creates an invalid and unselectable ID. In my case, my layer had the ID `#Above-Ground Tunnels _2`.

The proposed change will replace all spaces and results in the correctly hyphenated `#Above-Ground-Tunnels_2`.