Prevent base radio buttons in one map affecting another
Remove name attribute from base radio buttons, add tests and example for
manual testing. Addresses #474.

Giving each base layer radio button a name attribute of "base" causes an
issue when more than one map with a layer switcher control is shown on a
page; selecting a base layer causes all other radio buttons with the
same name to be unchecked. As we're not relying on the default radio
group behaviour due to the panel being re-rendered when a layer's
visibility changes this commit omits the name attribute.